Refactor Hyperopt-list and hyperopt-show to reduce some duplicate code
